Merge 154145 - When registering in HKLM, register everything in HKLM.
Don't mix-and-match registrations as it confuses Win7->Win8 upgrade machinery and we lose default browser if we have it...

This fixes a regression introduced in http://crrev.com/133333 on April 24th 2012. While registering RegisteredApplications in HKLM and the ProgId it points to in HKCU for user-level installs works (i.e. we are able to become default browser) on Win7, the Win7->Win8 OS upgrade doesn't understand this and we lose default when upgrading to Win8...

BUG=144910
TEST=Install Chrome on Win7, make Chrome default, update to Windows 8, and make sure Chrome is still default.

Review URL: https://chromiumcodereview.appspot.com/10899043

TBR=gab@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10910118

git-svn-id: http://src.chromium.org/svn/branches/1229/src@155262 4ff67af0-8c30-449e-8e8b-ad334ec8d88c