Merge 10912101 - Fix crash in extension NavigationObserver
This file was renamed since the R22 branch point. This is the same
code change as https://chromiumcodereview.appspot.com/10912101/ but
in a different file.

We were derefering a scoped_ptr<> immediately after resetting it.
This caused crashes on Linux and a top-crash on Chrome OS. Fixed
by copying out the data we need before we delete the object.

BUG=145741
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10928042

git-svn-id: http://src.chromium.org/svn/branches/1229/src@155268 4ff67af0-8c30-449e-8e8b-ad334ec8d88c