Merge 154352 - Force COM machinery to notice that the path to the DelegateExecute verb handler has changed.
On update, the path to the local COM server hosting the component changes.  Although the installer updates the path in the registry, it seems that the COM service that runs the handler doesn't notice this.  As a result, Chrome shortcuts stop working after updates.  This change forces COM to notice that the path has changed by attempting to instantiate the class after deleting the registration.  This forces COM to flush its cache, making subsequent uses of the handler by the shell succeed.

BUG=143338
TEST=see bug
TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10916018

TBR=grt@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10911139

git-svn-id: http://src.chromium.org/svn/branches/1229/src@155298 4ff67af0-8c30-449e-8e8b-ad334ec8d88c