Merge 151356 - [Sync] Dont upload an unrecoverable error for missing synced bookmark node
We were uploading when failing to associate the synced bookmarks node,
regardless of whether we expected it or not. Now we do the upload
at a layer above, where we know whether we care or not.

BUG=142387
TEST=Signing into sync on an account with no mobile devices should not
trigger an unrecoverable error upload.

Review URL: https://chromiumcodereview.appspot.com/10825325

TBR=zea@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10907126

git-svn-id: http://src.chromium.org/svn/branches/1229/src@155520 4ff67af0-8c30-449e-8e8b-ad334ec8d88c